feat: add lockfiles as they are automatically generated on install
BREAKING CHANGE: only needs to run once after ci
