+++ lib/update-lockfile.js
@@ -2,53 +2,15 @@
 
 const exec = require('child_process').execSync
 const _ = require('lodash')
-const semver = require('semver')
 
-const flags = {
-  'dependencies': ' -S',
-  'devDependencies': ' -D',
-  'optionalDependencies': ' -O'
-}
-
-const yarnFlags = {
-  'dependencies': '',
-  'devDependencies': ' -D',
-  'optionalDependencies': ' -O'
-}
-
-module.exports.updateLockfile = function updateLockfile (dependency, options) {
-  if (!options.yarn && semver.lt(exec('npm --version').toString().trim(), '3.0.0')) {
-    exec('npm shrinkwrap')
-  } else {
-    if (options.yarn) {
-      const flag = yarnFlags[dependency.type]
-      const envArgs = process.env.GK_LOCK_YARN_OPTS ? ` ${process.env.GK_LOCK_YARN_OPTS.trim()}` : ''
-      const args = `${flag}${envArgs} '${dependency.name}@${dependency.range}'`
-      exec(`yarn add${args}`)
-    }
-
-    if (options.npm) {
-      const flag = flags[dependency.type]
-      const prefix = dependency.prefix ? ` --save-prefix="${dependency.prefix}"` : ''
-      const args = `${flag}${prefix} ${dependency.name}@${dependency.version}`
-      var npmBin = 'npm'
-      try {
-        exec('npm5 -v')
-        npmBin = 'npm5'
-      } catch (err) {}
-      exec(`${npmBin} install${args}`)
-    }
-  }
-}
-
-module.exports.stageLockfile = function stageLockfile () {
+module.exports.stageLockfiles = function stageLockfiles () {
   // make sure that we have changes to add
   if (exec('git status --porcelain').toString() === '') return
 
   // stage the updated lockfile
-  exec('git add npm-shrinkwrap.json 2>/dev/null || true')
-  exec('git add package-lock.json 2>/dev/null || true')
-  exec('git add yarn.lock 2>/dev/null || true')
+  exec('git add ./**/npm-shrinkwrap.json 2>/dev/null || true')
+  exec('git add ./**/package-lock.json 2>/dev/null || true')
+  exec('git add ./**/yarn.lock 2>/dev/null || true')
 }
 
 module.exports.commitLockfiles = function commitLockfiles () {
